ucs16: exclusao (#21)
* feat: exclusao de fornecedor

* feat: fluxo alternativo 1 do caso de uso

* feat: fluxo alternativo 4 do caso de uso

* feat: fluxo alternativo 5

+++ KnnSystem/src/main/java/com/knnsystem/api/model/repository/FornecedorRepository.java
@@ -8,7 +8,7 @@ import com.knnsystem.api.model.entity.Fornecedor;
 import java.util.List;
 import java.util.Optional;
 
-public interface FornecedorRepository extends JpaRepository<Fornecedor, Integer> {
+public interface FornecedorRepository extends JpaRepository<Fornecedor, Long> {
 
     Optional<Fornecedor> findByCnpj(@CNPJ String cnpj);
 
@@ -22,4 +22,5 @@ public interface FornecedorRepository extends JpaRepository<Fornecedor, Integer>
     Optional<Fornecedor> findByNumControle(Long numControle);
 
     Optional<Fornecedor> findByRazaoSocial(String razaoSocial);
+
 }
